+++ joola.io.js
@@ -9,72 +9,6 @@
  *  @license GPL-3.0+ <http://spdx.org/licenses/GPL-3.0+>
  */
 
-var
-  uuid = require('node-uuid');
-
-//Dispatch = require('./lib/dispatch');
-
-require('./lib/common/globals');
-
-//our base global object
-var joola = {};
-global.joola = module.exports = joola;
-joola.VERSION = require('./package.json').version;
-joola.UID = uuid.v4();
-
-//setup the process domain, to handle unhandled exceptions
-require('./lib/common/domain');
-
-//setup the stack
-joola.common = require('./lib/common/index');
-joola.state = require('./lib/common/state');
-joola.config = require('./lib/common/config')({});
-joola.logger = require('./lib/common/logger');
-joola.events = require('./lib/common/events');
-joola.dispatch = require('./lib/dispatch')({});
-joola.sdk = require('./lib/sdk');
-joola.webserver = require('./lib/webserver');
-joola.stats = null; //require('./lib/common/stats');
-
-//we're live
-joola.logger.info('joola.io version ' + joola.VERSION + ' started.');
-
-//take the stack online
-joola.dispatch.hook();
-joola.redis = joola.config.stores.redis.redis;
-
-joola.webserver.start({}, function (err) {
-  if (err) {
-    joola.logger.error('Webserver startup reported an issue, exiting: ' + err);
-    return shutdown(1);
-  }
-  if (joola.config.get('webserver')) {
-    setTimeout(function () {
-      joola.logger.debug('Verifying webserver(s)');
-      if (joola.state.controls['webserver-http'].state != 'working' ||
-        (joola.state.controls['webserver-https'] && joola.state.controls['webserver-https'].state != 'working')) {
-        joola.logger.error('Failed to validate webserver(s) are online.');
-        return shutdown(1);
-      }
-      else
-        joola.logger.debug('Webserver(s) verified.');
-    }, 3000);
-  }
-});
-
-//setup the main watchers for system health
-require('./lib/common/watchers');
-
-//setup REPL for remote management
-// we have a special file that contains a set of global functions designed to be used by REPL.
-// NOTE: this option should be only used for debug/development and can be used as a back-door.
-if (joola.config.get('repl'))
-  require('./lib/common/repl');
-
-//load command line and env variables needed for startup.
-var cli = require('./lib/common/cli');
-if (cli.process()) {
-  //the user's args included some terminating ones, such as: version, help
-  process.exit(0);
-}
-joola.events.emit('init:done', 1);
+module.exports = process.env.NODE_COV
+  ? require('./lib-cov/joola.io')
+  : require('./lib/joola.io');
