+++ lib/common/domain.js
@@ -12,9 +12,8 @@
 var
   domain = require('domain');
 
-if (process.env.NODE_ENV == 'test')
-  return;
-
+if (process.env.NODE_ENV != 'test')
+{
 joola.domain = process.domain = domain.create();
 joola.domain.on('error', function (domain, err) {
   console.log('ERROR! ' + domain, err);
@@ -38,3 +37,4 @@ process.on('exit', function () {
 process.on('SIGINT', function () {
   global.shutdown(0);
 });
+  }
